update addToCart in home, index, product, product-detail

+++ js/index.js
@@ -94,12 +94,28 @@ function addToCart(id) {
   if(localStorage.getItem('cart') === null) {
     var idArray = new Array(id);
     localStorage.setItem('cart', JSON.stringify(idArray));
-    alert('ok');
+    toastr["success"]('<strong>Add to cart success.</strong>');
   }
   else {
-    idArray = JSON.parse(localStorage.getItem('cart'));
-    idArray.push (id);
-    localStorage.setItem('cart', JSON.stringify(idArray));
-    alert('ok');
+      var isExistP = checkExistInCart(id);
+
+      if (isExistP === false) {
+        idArray = JSON.parse(localStorage.getItem('cart'));
+        idArray.push(id);
+        localStorage.setItem('cart', JSON.stringify(idArray));
+        toastr["success"]('<strong>Add to cart success.</strong>');
+      }
   }
+}
+function checkExistInCart(id){
+  var isExist = false;
+  for(var i = 0; i < JSON.parse(localStorage.getItem('cart')).length; i++){
+    if((JSON.parse(localStorage.getItem('cart')))[i] === id){
+      toastr["success"]('<strong>Products already in the cart.</strong>');
+      isExist = true;
+      break;
+    }
+    isExist = false;
+  }
+  return isExist;
 }
